Add MainContent component and styles.
This commit adds the MainContent component and the necessary supporting code. This includes a new SCSS file for MainContent, a new SVG deleted, and updates to other components to make sure the MainContent component renders correctly. The changes also involve making updates to package-lock.json file to accommodate new dependencies.